split flrw protocol into pure and standard
