+++ src/cosmology/api/compat/__init__.py
@@ -3,8 +3,10 @@
 # LOCAL
 from cosmology.api.compat.core import CosmologyWrapperAPI
 from cosmology.api.compat.flrw import FLRWCosmologyWrapperAPI
+from cosmology.api.compat.flrwstandard import StandardFLRWCosmologyWrapperAPI
 
 __all__ = [
     "CosmologyWrapperAPI",
     "FLRWCosmologyWrapperAPI",
+    "StandardFLRWCosmologyWrapperAPI",
 ]
